Add upload image, add view images, add view reservations for property owner

+++ HakunaMatata/HakunaMatata.Application/CommandsHandlers/DeleteImageCommandHandler.cs
@@ -1,6 +1,7 @@
 ﻿using HakunaMatata.Application.Commands;
+using HakunaMatata.Application.Exceptions;
 using HakunaMatata.Core.Abstractions;
-using HakunaMatata.Core.Models;
+using HakunaMatata.Data.Services;
 using MediatR;
 using System;
 using System.Collections.Generic;
@@ -10,21 +11,36 @@ using System.Threading.Tasks;
 
 namespace HakunaMatata.Application.CommandsHandlers
 {
-    public class DeleteImageCommandHandler : IRequestHandler<DeleteImageCommand, Image>
+    public class DeleteImageCommandHandler : IRequestHandler<DeleteImageCommand>
     {
         private IUnitOfWork _uow;
+        private ITokenService _tokenService;
 
-        public DeleteImageCommandHandler(IUnitOfWork uow)
+        public DeleteImageCommandHandler(IUnitOfWork uow, ITokenService tokenService)
         {
             _uow = uow;
+            _tokenService = tokenService;
         }
 
-        public async Task<Image> Handle(DeleteImageCommand request, CancellationToken cancellationToken)
+        public async Task<Unit> Handle(DeleteImageCommand request, CancellationToken cancellationToken)
         {
-            var image = _uow.ImageRepository.DeleteById(request.ImageId);
-            await _uow.SaveAsync();
+            var userId = _tokenService.DecodeToken(request.Token);
+            var user = await _uow.UserRepository.GetByIdAsync(userId);
 
-            return image;
+            if (user.Property == null)
+                throw new UserDoesNotHaveProperty("User doesn't have a property assigned");
+
+            foreach (var image in user.Property.Images)
+            {
+                if (image.ImageId == request.ImageId)
+                {
+                    _uow.ImageRepository.DeleteById(request.ImageId);
+                    await _uow.SaveAsync();
+                    return new Unit();
+                }
+            }
+
+            throw new PropertyDoesNotHaveThisImageException("This image isn't from you property");
         }
     }
 }
